mcp-server/pyproject.toml
[project]
name = "memory-kit-mcp"
version = "0.8.0"
description = "Persistent Markdown vault MCP server for SecondBrain"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "SI-GMT", email = "imagerienumerique@gmail.com" },
]
keywords = ["mcp", "memory", "vault", "markdown", "obsidian", "secondbrain"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
]
dependencies = [
    "fastmcp>=2.13",
    "pydantic>=2.9",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/SI-GMT/SecondBrain"
Repository = "https://github.com/SI-GMT/SecondBrain"
Issues = "https://github.com/SI-GMT/SecondBrain/issues"

[project.scripts]
memory-kit-mcp = "memory_kit_mcp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5",
    "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/memory_kit_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=memory_kit_mcp --cov-report=term-missing --cov-fail-under=80"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]  # line length handled by formatter
